chore: make RetryContext an interface to allow for decoration
